Implement primitive layer toggle mechanism
Due to the complexity/unfamiliarity of/with the transient package, the mechanism
here is very basic (and cumbersome). After a layer is added, it should first be
activated by pressing `L`. Then to render a list containing the numbers of the
to be rendered layers must be set via pressing `-L`.